+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.goshawk 0.1.14.9007
+# teal.goshawk 0.1.14.9008
 
 # Examples now use `scda.2022` data instead of `scda.2021`.